Path Builder allow to build path from components A root path, optional prefixes and a suffix (path inside the root) It manages the path creation for the global out path Path can be created from several components with the general form root / [prefixes] / suffix prefixes are optional The Builder can also use an absolute mode, using directly a full path, in this mode the components are not used

Path Builder allow to build path from components A root path, optional prefixes and a suffix (path inside the root) It manages the path creation for the global out path Path can be created from several components with the general form root / [prefixes] / suffix prefixes are optional The Builder can also use an absolute mode, using directly a full path, in this mode the components are not used

Path Builder allow to build path from components A root path, optional prefixes and a suffix (path inside the root) It manages the path creation for the global out path Path can be created from several components with the general form root / [prefixes] / suffix prefixes are optional The Builder can also use an absolute mode, using directly a full path, in this mode the components are not used

Methods

Public methods

PathBuilder$new()

PathBuilder$set_root()

PathBuilder$get_root()

PathBuilder$update()

PathBuilder$set_suffix()

PathBuilder$get_suffix()

PathBuilder$set_full_path()

PathBuilder$path()

PathBuilder$set_prefixes()

PathBuilder$set_prefix()

PathBuilder$get_prefixes()

PathBuilder$is_absolute()

PathBuilder$components()

PathBuilder$clone()

Method new()

Create the path builder instance for the given root path

Usage

PathBuilder$new(root)

Arguments

root

character root path

Method set_root()

Set the root path of the builder

Usage

PathBuilder$set_root(root)

Arguments

root

the root path to set

Method get_root()

Get the root pah

Usage

PathBuilder$get_root()

Method update()

Rebuilt current path from the path components If absolute mode is on, the full path will be returned, otherwise the path will be rebuild from components

Usage

PathBuilder$update()

Method set_suffix()

Define the current suffix component of the path

Usage

PathBuilder$set_suffix(path)

Arguments

path

character

Method get_suffix()

Get the suffix component of the path

Usage

PathBuilder$get_suffix()

Method set_full_path()

Define the full path

If path is not null, the full path is defined by the provided one, ignoring other paths components (root, prefixes, suffix) If provided path is null, then absolute mode is disabled and the path is rebuild from components

Usage

PathBuilder$set_full_path(path)

Arguments

path

full path to used

Method path()

Get a path inside the current path

Usage

PathBuilder$path(...)

Arguments

...

character arguments to concat as sub path

Method set_prefixes()

Set the list of the prefixes

Prefixes are named components. Each prefix entry is added when the path is created between the root and the suffix root / prefixes... / suffix This can be used to add subpath before suffix without changing the root (to create the same tree layout in different root's subpath for example) Prefixes are used in order of the list

Usage

PathBuilder$set_prefixes(prefixes)

Arguments

prefixes

list()

Method set_prefix()

Set a prefix with a given name

Usage

PathBuilder$set_prefix(name, value)

Arguments

name

character name of the prefix

value

value to use for the named prefix

Method get_prefixes()

Get prefixes

Usage

PathBuilder$get_prefixes()

Method is_absolute()

Is the current path using absolute mode If absolute mode is set, the path is set directly as a full path, and other path components are not used

Usage

PathBuilder$is_absolute()

Method components()

Export components as a static list

Usage

PathBuilder$components()

Method clone()

The objects of this class are cloneable with this method.

Usage

PathBuilder$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.