Register a path prefix for the global out path

Each path prefix will be added to the root path separated by '/' to create a full path before the suffix.

Usage

add_out_path_prefix(name, prefix)

Arguments

name

name of the prefix

prefix

path to add to the prefix

Details

Prefix mechanism allows to define constant path at several levels. For example : output path for global out path output path for a project (that will complete base.out.path) output path in a script of the project

See also

Examples

if (FALSE) { # \dontrun{
set_root_out_path('/my/path') # root of the out path is /my/path
add_path_prefix('project', 'my_project')
init_path('output') # Add output to the current path as suffix
} # }
my_path() # => '/my/path/my_project/output'
#> [1] "/tmp/RtmpD6jaBz/R.INSTALL1b5257f7fe72/workspace/"