Define subpath (suffix) of global out path, usable by my_path()

init_path() and my_path() are dedicated to manage path to files used by scripts to make them independent from the actual physical location of the files, which depends on where (on which machine, account) the script is running.

Usage

init_path(p, full.path = FALSE, create = TRUE)

Arguments

p

character path

full.path

logical if TRUE p is considered as an absolute path, replace all the current path

create

logical, if TRUE create the subpath if not exists

Details

By default the path is added as a suffix to the global output path (see set_root_out_path ), unless the parameter full.path is TRUE

See also