Restore the dev environment from a restore point

This function restores the development environment from a previously created restore point, allowing you to revert your workspace to a previous state.

workspace_restore(n_threads = 2, is_interactive = interactive())

Arguments

n_threads

Integer. Default 2. The number of threads to use for the saving process.

is_interactive

Logical. Default interactive(). This parameter is used for package development purposes only, and it is recommended to avoid changing this setting.

Value

Nothing.

Details

The workspace_restore() function restores the current development environment from a save state. When using this function two options will be presented:

Latest

Selecting the latest restore point issues a warning about resetting the global environment and asks for confirmation

Upon confirming, it restores the workspace to the latest restore point and displays a "Restore complete" message

List out the restore points

Selecting to list out the restore points presents a menu with available restore points

Including the date and time along with any optional note associated with each restore point

Upon selecting a specific restore point, it warns about resetting the global environment and asks for confirmation

Upon confirming, it restores the workspace to the chosen restore point and displays a "Restore complete" message

See also

Examples

if (FALSE) {
# Initiate the restore process
workspace_restore()
}